Change conp matrix to independent class
Apply same changes to matrix as for ConpVector in previous commit.
ConpVector and ConpMatrix both get tag_to_iele from setup in fix
instead of calculating multiple times.
